<!-- Thank you for contributing! -->

### Description

#65 

<!-- Please insert your description here and provide especially info about the "what" this PR is solving -->

### Additional context

Use recursion to replace flat.

<!-- e.g. is there anything you'd like reviewers to focus on? -->

---

### What is the purpose of this pull request?

Just code golfing performance, inspired by @TIMMLOPK

Original
```
CVA x 931,787 ops/sec ±0.62% (89 runs sampled)
CVA with tailwind-merge x 503,360 ops/sec ±1.16% (88 runs sampled)
TV with slots x 97,211 ops/sec ±0.44% (91 runs sampled)
TV without slots x 113,086 ops/sec ±0.57% (90 runs sampled)
Fastest is CVA
```

@TIMMLOPK's branch
```
CVA x 957,529 ops/sec ±0.54% (90 runs sampled)
CVA with tailwind-merge x 475,375 ops/sec ±1.19% (86 runs sampled)
TV with slots x 151,382 ops/sec ±3.36% (86 runs sampled)
TV without slots x 180,776 ops/sec ±0.53% (88 runs sampled)
Fastest is CVA
```

This branch
```
CVA x 940,169 ops/sec ±0.57% (91 runs sampled)
CVA with tailwind-merge x 489,335 ops/sec ±0.79% (90 runs sampled)
TV with slots x 181,157 ops/sec ±0.72% (89 runs sampled)
TV without slots x 210,579 ops/sec ±0.60% (91 runs sampled)
Fastest is CVA
```

<!-- (put an "X" next to an item) -->

- [ ] Bug fix
- [ ] New Feature
- [ ] Documentation update
- [X] Other

### Before submitting the PR, please make sure you do the following

- [X] Read the [Contributing Guidelines](https://github.com/jrgarciadev/tailwind-variants/blob/main/CONTRIBUTING.md).
- [X] Follow the [Style Guide](https://github.com/jrgarciadev/tailwind-variants/blob/main/CONTRIBUTING.md#style-guide).
- [X] Check that there isn't already a PR that solves the problem the same way to avoid creating a duplicate.
- [X] Provide a description in this PR that addresses **what** the PR is solving, or reference the issue that it solves (e.g. `fixes #123`).